Decouple conformance tests into modular files for Python and Kotlin SDKs
